+++ package/pascaltzreg.pas
@@ -1,0 +1,19 @@
+unit PascalTZReg;
+
+{$mode objfpc}{$H+}
+
+interface
+
+procedure Register;
+
+implementation
+
+uses
+  Classes, LResources, uPascalTZ;
+
+procedure Register;
+begin
+  RegisterComponents('Misc', [TPascalTZ]);
+end;
+
+end.